Renaming the DN database and moving it to DN package
