+++ core/src/main/java/io/plaidapp/core/designernews/DesignerNewsInjection.kt
@@ -23,10 +23,12 @@ import com.google.gson.Gson
 import com.jakewharton.retrofit2.adapter.kotlin.coroutines.experimental.CoroutineCallAdapterFactory
 import io.plaidapp.core.BuildConfig
 import io.plaidapp.core.data.api.DenvelopingConverter
+import io.plaidapp.core.designernews.data.LoggedInUserDao
 import io.plaidapp.core.designernews.data.api.ClientAuthInterceptor
 import io.plaidapp.core.designernews.data.api.DesignerNewsService
 import io.plaidapp.core.designernews.data.comments.CommentsRemoteDataSource
 import io.plaidapp.core.designernews.data.comments.CommentsRepository
+import io.plaidapp.core.designernews.data.database.DesignerNewsDatabase
 import io.plaidapp.core.designernews.data.login.AuthTokenLocalDataSource
 import io.plaidapp.core.designernews.data.login.LoginLocalDataSource
 import io.plaidapp.core.designernews.data.login.LoginRemoteDataSource
@@ -54,6 +56,10 @@ fun provideLoginLocalDataSource(context: Context): LoginLocalDataSource {
         LoginLocalDataSource.DESIGNER_NEWS_PREF
     )
     return LoginLocalDataSource(preferences)
+}
+
+fun provideLoggedInUserDao(context: Context): LoggedInUserDao {
+    return DesignerNewsDatabase.getInstance(context).loggedInUserDao()
 }
 
 fun provideLoginRepository(context: Context): LoginRepository {
