+++ core/src/main/java/io/plaidapp/core/designernews/data/database/DesignerNewsDatabase.kt
@@ -14,13 +14,14 @@
  * limitations under the License.
  */
 
-package io.plaidapp.core.data
+package io.plaidapp.core.designernews.data.database
 
 import android.arch.persistence.room.Database
 import android.arch.persistence.room.Room
 import android.arch.persistence.room.RoomDatabase
 import android.arch.persistence.room.TypeConverters
 import android.content.Context
+import io.plaidapp.core.data.Converters
 import io.plaidapp.core.designernews.data.LoggedInUserDao
 import io.plaidapp.core.designernews.data.users.model.LoggedInUser
 
@@ -29,7 +30,7 @@ import io.plaidapp.core.designernews.data.users.model.LoggedInUser
  */
 @Database(entities = [LoggedInUser::class], version = 1, exportSchema = false)
 @TypeConverters(Converters::class)
-abstract class AppDatabase : RoomDatabase() {
+abstract class DesignerNewsDatabase : RoomDatabase() {
     abstract fun loggedInUserDao(): LoggedInUserDao
 
     companion object {
@@ -37,16 +38,22 @@ abstract class AppDatabase : RoomDatabase() {
         private const val DATABASE_NAME = "plaid-db"
 
         // For Singleton instantiation
-        @Volatile private var instance: AppDatabase? = null
+        @Volatile private var instance: DesignerNewsDatabase? = null
 
-        fun getInstance(context: Context): AppDatabase {
-            return instance ?: synchronized(this) {
-                instance ?: buildDatabase(context).also { instance = it }
+        fun getInstance(context: Context): DesignerNewsDatabase {
+            return instance
+                ?: synchronized(this) {
+                instance
+                    ?: buildDatabase(
+                        context
+                    ).also { instance = it }
             }
         }
 
-        private fun buildDatabase(context: Context): AppDatabase {
-            return Room.databaseBuilder(context, AppDatabase::class.java, DATABASE_NAME).build()
+        private fun buildDatabase(context: Context): DesignerNewsDatabase {
+            return Room.databaseBuilder(context, DesignerNewsDatabase::class.java,
+                DATABASE_NAME
+            ).build()
         }
     }
 }
